+++ R/Swrappers.R
@@ -96,8 +96,8 @@ alter <- function(x, ...) {
 }
 #' Alter Time Series
 #'
-#' Shortens, lengthens and/or changes the periodicity of a \code{\link{DTSg}}
-#'  object.
+#' Shortens (subsets), lengthens and/or changes the periodicity of a
+#'  \code{\link{DTSg}} object.
 #'
 #' @param x A \code{\link{DTSg}} object (S3 method only).
 #' @param from A \code{\link{POSIXct}} date with the same time zone as the time
@@ -125,7 +125,7 @@ alter <- function(x, ...) {
 #' # new DTSg object
 #' x <- DTSg$new(values = flow)
 #'
-#' # extract the first two years
+#' # subset the first two years
 #' ## R6 method
 #' x$alter(from = "2007-01-01", to = "2008-12-31")
 #'
@@ -195,6 +195,14 @@ colapply <- function(x, ...) {
 #' @param cols A character vector specifying the columns to apply \code{fun} to.
 #' @param clone A logical specifying if the object is modified in place or if a
 #'  clone (copy) is made beforehand.
+#' @param resultCols An optional character vector of the same length as
+#'  \code{cols}. Non-existing columns specified in this argument are added and
+#'  existing columns are overwritten by the return values of \code{fun}. Columns
+#'  are matched elementwise between \code{resultCols} and \code{cols}.
+#' @param suffix An optional character string. The return values of \code{fun}
+#'  are added as new columns with names consisting of the columns specified in
+#'  \code{cols} and this suffix. Existing columns are never overwritten. Only
+#'  used when \code{resultCols} is not specified.
 #'
 #' @details
 #' In addition to the \code{\dots} argument, this method hands over a
@@ -240,18 +248,19 @@ cols <- function(x, ...) {
 }
 #' Get Column Names
 #'
-#' Queries all column names of a \code{\link{DTSg}} object or those of a certain
-#'  \code{\link{class}} only.
+#' Queries all column names of a \code{\link{DTSg}} object, those of certain
+#'  \code{\link{class}}es and/or those matching a certain pattern only.
 #'
 #' @param x A \code{\link{DTSg}} object (S3 method only).
-#' @param class A character string matched to the most specific class (first
-#'  element) of each column's \code{\link{class}} vector or \code{"all"} for all
-#'  column names.
-#' @param \dots Not used (S3 method only).
+#' @param class An optional character vector matched to the most specific class
+#'  (first element) of each column's \code{\link{class}} vector.
+#' @param pattern An optional character string passed on to the \code{pattern}
+#'  argument of \code{\link{grep}}.
+#' @param \dots Further arguments passed on to \code{\link{grep}}.
 #'
 #' @return Returns a character vector.
 #'
-#' @seealso \code{\link{DTSg}}, \code{\link{class}}
+#' @seealso \code{\link{DTSg}}, \code{\link{class}}, \code{\link{grep}}
 #'
 #' @examples
 #' # new DTSg object
@@ -273,7 +282,7 @@ cols.DTSg <- S3WrapperGenerator(expression(DTSg$public_methods$cols))
 #' Merge Two DTSg Objects
 #'
 #' Joins two \code{\link{DTSg}} objects based on their \emph{.dateTime} column.
-#'  Their time zones and \emph{aggregated} field must be the same.
+#'  Their time zones and \emph{aggregated} fields must be the same.
 #'
 #' @param x A \code{\link{DTSg}} object (S3 method only).
 #' @param y A \code{\link{DTSg}} object or an object coercible to one. See
@@ -400,10 +409,11 @@ setMethod(
 #'  up to this date.
 #' @param cols A character vector specifying the columns whose values shall be
 #'  plotted.
-#' @param secAxisCols A character vector specifying the columns whose values
-#'  shall be plotted on a secondary axis. Must be a subset of \code{cols}.
-#' @param secAxisLabel A character string specifying the label of the secondary
-#'  axis.
+#' @param secAxisCols An optional character vector specifying the columns whose
+#'  values shall be plotted on a secondary axis. Must be a subset of
+#'  \code{cols}.
+#' @param secAxisLabel A character string specifying the label of the optional
+#'  secondary axis.
 #' @param \dots Not used (S3 method only).
 #'
 #' @return Returns a \code{\link{DTSg}} object.
@@ -480,10 +490,10 @@ rollapply <- function(x, ...) {
 #' @param \dots Further arguments passed on to \code{fun}.
 #' @param cols A character vector specifying the columns whose rolling window
 #'  \code{fun} shall be applied to.
-#' @param before A numeric specifying the size of the window in time steps
-#'  before the \dQuote{center} of the rolling window.
-#' @param after A numeric specifying the size of the window in time steps after
-#'  the \dQuote{center} of the rolling window.
+#' @param before An integerish value specifying the size of the window in time
+#'  steps before the \dQuote{center} of the rolling window.
+#' @param after An integerish value specifying the size of the window in time
+#'  steps after the \dQuote{center} of the rolling window.
 #' @param weights A character string specifying a method to calculate weights
 #'  for \code{fun}, for instance, \code{\link{weighted.mean}}. See details for
 #'  further information.
@@ -491,6 +501,14 @@ rollapply <- function(x, ...) {
 #'  \code{weights}. See details for further information.
 #' @param clone A logical specifying if the object is modified in place or if a
 #'  clone (copy) is made beforehand.
+#' @param resultCols An optional character vector of the same length as
+#'  \code{cols}. Non-existing columns specified in this argument are added and
+#'  existing columns are overwritten by the return values of \code{fun}. Columns
+#'  are matched elementwise between \code{resultCols} and \code{cols}.
+#' @param suffix An optional character string. The return values of \code{fun}
+#'  are added as new columns with names consisting of the columns specified in
+#'  \code{cols} and this suffix. Existing columns are never overwritten. Only
+#'  used when \code{resultCols} is not specified.
 #'
 #' @details
 #' In addition to the \code{\dots} argument, this method hands over the weights
@@ -581,11 +599,12 @@ values <- function(x, ...) {
 #'  reference to the \emph{values} is returned. See details for further
 #'  information.
 #' @param drop A logical specifying if the object and all references to it shall
-#'  be removed from the global environment after successfully querying its
-#'  values. This feature allows for a ressource efficient destruction of a
-#'  \code{\link{DTSg}} object while preserving its \emph{values.}
+#'  be removed from the global (and only the global) environment after
+#'  successfully querying its values. This feature allows for a ressource
+#'  efficient destruction of a \code{\link{DTSg}} object while preserving its
+#'  \emph{values.}
 #' @param class A character string specifying the class of the returned
-#'  \emph{values.} \code{"data.frame"} only works if either a copy of the
+#'  \emph{values.} \code{"data.frame"} only works when either a copy of the
 #'  \emph{values} is returned or the object is dropped.
 #' @param \dots Not used (S3 method only).
 #'