+++ R/aaa.R
@@ -1,7 +1,4 @@
-#' @import assertive.base
-#' @import assertive.numbers
-#' @import assertive.sets
-#' @import assertive.types
+#' @import checkmate
 #' @import data.table
 #' @import methods
 #' @import R6