+++ R/funinterpolate.R
@@ -1,3 +1,11 @@
+# no R CMD check notes
+.dateTime <- NULL
+.colAfter <- NULL
+.colBefore <- NULL
+.dateTimeAfter <- NULL
+.dateTimeBefore <- NULL
+.divisor <- NULL
+
 #' Linear Interpolation
 #'
 #' Linearly interpolates missing values of a numeric vector. For use with the
@@ -38,8 +46,8 @@
 #'
 #' @export
 interpolateLinear <- function(.col, roll = Inf, rollends = TRUE, .helpers) {
-  assert_is_numeric(.col)
-  assert_all_are_greater_than(assert_is_a_number(roll), 0)
+  qassert(.col, "n+")
+  qassert(roll, "N1(0,]")
 
   if (.helpers$periodicity != "unrecognised") {
     roll <- roll * as.numeric(.helpers$maxLag, units = "secs")
@@ -49,14 +57,6 @@ interpolateLinear <- function(.col, roll = Inf, rollends = TRUE, .helpers) {
   values <- data.table(.dateTime = .helpers$.dateTime, .col = .col, key = ".dateTime")
   values <- values[!is.na(.col), ]
 
-  # no R CMD check notes
-  .dateTime <- NULL
-  .colAfter <- NULL
-  .colBefore <- NULL
-  .dateTimeAfter <- NULL
-  .dateTimeBefore <- NULL
-  .divisor <- NULL
-
   DT <- values[
     DT,
     c(".dateTime", sprintf("x.%s", c(".dateTime", ".col"))),
@@ -64,7 +64,7 @@ interpolateLinear <- function(.col, roll = Inf, rollends = TRUE, .helpers) {
     roll = roll,
     rollends = rollends
   ]
-  setnames(DT, 2L:3L, c(".dateTimeBefore", ".colBefore"))
+  setnames(DT, 2:3, c(".dateTimeBefore", ".colBefore"))
 
   DT <- values[
     DT,
@@ -73,7 +73,7 @@ interpolateLinear <- function(.col, roll = Inf, rollends = TRUE, .helpers) {
     roll = -roll,
     rollends = rollends
   ]
-  setnames(DT, 4L:5L, c(".dateTimeAfter", ".colAfter"))
+  setnames(DT, 4:5, c(".dateTimeAfter", ".colAfter"))
 
   DT[, .divisor := as.numeric(.dateTimeAfter - .dateTimeBefore, units = "secs")]
   DT[, .col := .colBefore]