+++ NEWS.md
@@ -1,6 +1,16 @@
+## DTSg v0.3.0
+
+* Added `resultCols` and `suffix` arguments to `colapply` and `rollapply` methods: allows for adding return values of applied functions as new columns instead of replacing existing ones
+* `class` argument of `cols` method now accepts a character vector of class names
+* Changed default value of `class` argument of `cols` method from `"all"` to `NULL`, however, for backward compatibility `"all"` can still be used for the same result, but will eventually be treated as a filter for classes of type `all`
+* Added `pattern` and `...` arguments to `cols` method: allows for searching column names
+* Switched to argument check functions of `checkmate` package instead of `assertive.base`, `assertive.numbers`, `assertive.sets` and `assertive.types` packages
+* Slightly improved vignettes and documentation
+* Minor internal code improvements
+
 ## DTSg v0.2.1
 
-* Fixed error "Unsupported type passed to argument 'data'." related to `dygraphs` in at least some previous releases of R ≤ 3.5.3)
+* Fixed error "Unsupported type passed to argument 'data'." related to `dygraphs` in at least some previous releases of R ≤ 3.5.3
 
 ## DTSg v0.2.0